requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
azure-cognitiveservices-speech==1.43.0
bitsandbytes==0.45.5
cut-cross-entropy==25.1.1
datasets==3.5.0
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
environs==14.1.1
fastapi==0.115.12
fastapi_cors==0.0.6
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gTTS==2.5.4
hf_transfer==0.1.9
huggingface-hub==0.30.2
Jinja2==3.1.6
jiter==0.9.0
llvmlite==0.44.0
numpy==2.1.3
pandas==2.2.
peft==0.15.1
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.1
pycairo==1.27.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==0.21.1
python-slugify==8.0.4
pytz==2025.2
regex==2024.11.6
safetensors==0.5.3
sentencepiece==0.2.0
setuptools==75.8.2
shtab==1.7.1
six==1.17.0
sox==1.5.0
stable-ts==2.18.3
starlette==0.46.1
sympy==1.13.1
text-unidecode==1.3
tiktoken==0.3.1
tokenizers==0.21.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
transformers==4.51.2
triton-windows==3.2.0.post17
trl==0.15.2
typeguard==4.4.2
typing_extensions==4.13.2
tyro==0.9.18
tzdata==2025.2
unsloth==2025.3.19
unsloth_zoo==2025.3.17
uvicorn==0.34.0
wheel==0.45.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.19.0
zstandard==0.23.0
